+++ src/UglyToad.PdfPig.Tests/Graphics/Operations/GraphicsStateOperationTests.cs
@@ -95,17 +95,17 @@
                 {
                     result[i] = NameToken.Create("Hog");
                 }
-                else if (type == typeof(decimal))
+                else if (type == typeof(double))
                 {
-                    result[i] = 0.5m;
+                    result[i] = 0.5;
                 }
                 else if (type == typeof(int))
                 {
                     result[i] = 1;
                 }
-                else if (type == typeof(decimal[]) || type == typeof(IReadOnlyList<decimal>))
+                else if (type == typeof(double[]) || type == typeof(IReadOnlyList<double>))
                 {
-                    result[i] = new decimal[]
+                    result[i] = new double[]
                     {
                         1, 0, 0, 1, 2, 5
                     };
